Add files via upload
This is the pre alpha version of a video and gif modification software.  New features such as extended support for other video formats and the possibility of selecting a specific part of the video to convert to gif is also being implemented. The code is free and open source, so please fell free to reuse it while crediting the original author(s)